start `blit-displays` decomp & renderer + improve decompilation of some DMA macros (#2616)
Adds sprite distort, fixes buggy sprite rendering in progress, adds
scissoring support (used in various scrolling menus) and a very basic
implementation of `blit-displays`. This is enough to make the fade
effect in the progress menu work, along with all the menus working
properly without needing to use the REPL. This does not make screen
flipping and the filter when failing a mission work.

Added support in the decompiler for detecting `dma-buffer-add-gs-set`
and `dma-buffer-add-gs-set-flusha` and updated all of the Jak 2 code to
use it. Readability improved!

Fixes decompiler issue with `with-dma-buffer-add-bucket` not inlining
forms which broke syntax. Fixes store error warnings showing up for
non-existent stores, there is now a dedicated pass for this at the end.

I started work on making `BITBLTBUF` stuff work in the DirectRenderer,
but stopped for now because it wasn't strictly necessary. It will still
assert like before.

+++ game/graphics/texture/TexturePool.h
@@ -12,6 +12,7 @@
 #include "common/util/SmallVector.h"
 #include "common/versions/versions.h"
 
+#include "game/graphics/pipelines/opengl.h"
 #include "game/graphics/texture/TextureConverter.h"
 
 // verify all texture lookups.
@@ -140,7 +141,7 @@ class TextureMap {
  * The lowest level reference to texture data.
  */
 struct TextureData {
-  u64 gl = -1;               // the OpenGL texture ID
+  GLuint gl = -1;            // the OpenGL texture ID
   const u8* data = nullptr;  // pointer to texture data (owned by the loader)
 };
 
@@ -196,7 +197,7 @@ struct GpuTexture {
  * source will be non-null and the gpu_texture will be a placeholder that is safe to use.
  */
 struct TextureVRAMReference {
-  u64 gpu_texture = -1;  // the OpenGL texture to use when rendering.
+  GLuint gpu_texture = -1;  // the OpenGL texture to use when rendering.
   GpuTexture* source = nullptr;
 };
 
@@ -209,7 +210,7 @@ struct TextureInput {
 
   PcTextureId id;
 
-  u64 gpu_texture = -1;
+  GLuint gpu_texture = -1;
   bool common = false;
   const u8* src_data;
   u16 w, h;
